(#1114) Unit test tidy ups. Black-box UT for 3d gridscan ispyb. Rationalisation of begin_deposition for gridscan
